pyproject.toml
[tool.poetry]
name = "task-plastilin-func"
version = "0.1.0"
description = ""
authors = ["Dmitrii, <dima12sik@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
